--- a/lib/html5ever.ex
+++ b/lib/html5ever.ex
@@ -37,12 +37,8 @@ defmodule Html5ever do
        ]}
 
   """
-  def parse(html) when byte_size(html) > 500 do
-    parse_async(html)
-  end
-
   def parse(html) do
-    parse_sync(html)
+    parse_dirty(html)
   end
 
   @doc """
@@ -96,15 +92,11 @@ defmodule Html5ever do
        }}
 
   """
-  def flat_parse(html) when byte_size(html) > 500 do
-    flat_parse_async(html)
-  end
-
   def flat_parse(html) do
-    flat_parse_sync(html)
+    flat_parse_dirty(html)
   end
 
-  defp parse_sync(html) do
+  defp parse_dirty(html) do
     case Html5ever.Native.parse_sync(html) do
       {:html5ever_nif_result, :ok, result} ->
         {:ok, result}
@@ -114,32 +106,8 @@ defmodule Html5ever do
     end
   end
 
-  defp parse_async(html) do
-    :ok = Html5ever.Native.parse_async(html)
-
-    receive do
-      {:html5ever_nif_result, :ok, result} ->
-        {:ok, result}
-
-      {:html5ever_nif_result, :error, err} ->
-        {:error, err}
-    end
-  end
-
-  defp flat_parse_sync(html) do
+  defp flat_parse_dirty(html) do
     case Html5ever.Native.flat_parse_sync(html) do
-      {:html5ever_nif_result, :ok, result} ->
-        {:ok, result}
-
-      {:html5ever_nif_result, :error, err} ->
-        {:error, err}
-    end
-  end
-
-  defp flat_parse_async(html) do
-    :ok = Html5ever.Native.flat_parse_async(html)
-
-    receive do
       {:html5ever_nif_result, :ok, result} ->
         {:ok, result}