--- a/lib/html5ever/native.ex
+++ b/lib/html5ever/native.ex
@@ -39,9 +39,9 @@ defmodule Html5ever.Native do
   use Rustler, opts
 
   def parse_sync(_binary), do: err()
-  def parse_async(_binary), do: err()
+  def parse_dirty(_binary), do: err()
   def flat_parse_sync(_binary), do: err()
-  def flat_parse_async(_binary), do: err()
+  def flat_parse_dirty(_binary), do: err()
 
   defp err, do: :erlang.nif_error(:nif_not_loaded)
 end