--- a/native/html5ever_nif/Cargo.toml
+++ b/native/html5ever_nif/Cargo.toml
@@ -18,3 +18,4 @@ markup5ever = "0.10"
 tendril = "0.4"
 lazy_static = "1.4"
 scoped-pool = "1.0"
+